Crud.

+++ src/main/java/com/repository/TemaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.model.Tema;
 
-public interface TemaRepository extends JpaRepository<Tema, Integer>{
+public interface TemaRepository extends JpaRepository<Tema, Long>{
 
 }
